put compilation tests in their own module
